Moved Proposal type to another file
